add ObjectId in Schema for relationship / bug fix => user-password have to change at change-password (add isModified in User model)

+++ src/controllers/videoController.js
@@ -1,3 +1,4 @@
+import User from "../models/User";
 import Video from "../models/Video";
 
 export const home = async (req, res) => {
@@ -7,7 +8,7 @@ export const home = async (req, res) => {
 
 export const watch = async (req, res) => {
   const { id } = req.params;
-  const video = await Video.findById(id);
+  const video = await Video.findById(id).populate("owner");
   if (!video) {
     return res.render("404", { pageTitle: "Video Not found." });
   }
@@ -15,10 +16,18 @@ export const watch = async (req, res) => {
 };
 
 export const getEdit = async (req, res) => {
-  const { id } = req.params;
+  const {
+    session: {
+      user: { _id },
+    },
+    parmas: { id },
+  } = req;
   const video = await Video.findById(id);
   if (!video) {
-    return res.render("404", { pageTitle: "Video Not found." });
+    return res.status(404).render("404", { pageTitle: "Video Not found." });
+  }
+  if (String(_id) !== String(video.owner)) {
+    return res.status(403).redirect("/");
   }
   res.render("edit", { pageTitle: `Edit: ${video.title}`, video });
 };
@@ -44,13 +53,24 @@ export const getUpload = (req, res) => {
 
 export const postUpload = async (req, res) => {
   //here we will add a video to the videos array.
-  const { title, description, hashtags } = req.body;
+  const {
+    session: {
+      user: { _id },
+    },
+    file: { path: fileUrl },
+    body: { title, description, hashtags },
+  } = req;
   try {
-    await Video.create({
+    const newVideo = await Video.create({
       title,
+      fileUrl,
+      owner: _id,
       description,
       hashtags: Video.formatHashtags(hashtags),
     });
+    const user = await User.findById(_id);
+    user.videos.push(newVideo._id);
+    user.save();
     return res.redirect("/");
   } catch (err) {
     return res.render("upload", {
@@ -61,7 +81,19 @@ export const postUpload = async (req, res) => {
 };
 
 export const deleteVideo = async (req, res) => {
-  const { id } = req.params;
+  const {
+    session: {
+      user: { _id },
+    },
+    params: { id },
+  } = req;
+  const video = await Video.findById(id);
+  if (!video) {
+    return res.status(404).render("404", { pageTitle: "Video Not found." });
+  }
+  if (String(_id) !== String(video.owner)) {
+    return res.status(403).redirect("/");
+  }
   await Video.findByIdAndDelete(id);
   return res.redirect("/");
 };
